requirements.txt
djangorestframework==3.9.0
gunicorn
django-heroku
django-rest-swagger
